requirements.txt
appdirs==1.4.4
art==5.2
certifi>=2024.0.0
cffi>=1.15.0
chardet==4.0.0
Deprecated==1.2.12
diskcache>=5.6.0
future==0.18.2
idna>=3.4
Jinja2>=3.1.0
Markdown>=3.5.0
MarkupSafe>=2.1.0
minify_html>=0.15.0
oauthlib>=3.2.0
pycparser==2.20
PyGithub>=2.1.0
Pygments>=2.17.0
PyJWT>=2.8.0
PyNaCl>=1.5.0
python-dotenv>=1.0.0
python-slugify>=8.0.0
python-twitter==3.5
rcssmin>=1.2.2
requests>=2.31.0
requests-oauthlib>=2.0.0
six==1.16.0
text-unidecode==1.3
tqdm>=4.66.0
urllib3>=2.0.0
wrapt==1.16.0
